==> 2_eda.ipynb <==
# Exploratory Data Analysis <a class="tocSkip">

# Loading useful extensions
%load_ext autoreload
%autoreload
%load_ext nb_black
%matplotlib inline

# Import packages
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

# Load Data

df_eda = pd.read_csv("df_eda.csv")
del df_eda["Unnamed: 0"]

# EDA

The data has 26 dimensions, which are listed below. Some of the attribute names are quite ambiguous, such are the codes for tax and building classes. For more information you can have a look at https://www1.nyc.gov/assets/finance/downloads/pdf/07pdf/glossary_rsf071607.pdf and https://www1.nyc.gov/assets/finance/jump/hlpbldgcode.html .

# Checking for NaN values
df_eda.isna().sum().sort_values(ascending=False)

# Dimension of the data
df_eda.shape

sum(df_eda.duplicated(df_eda.columns))

sns.distplot(df_eda["price"])
# sns.distplot(df_eda["price"]).get_figure().savefig("output.png", dpi=600)

sales = np.log(df_eda["price"])
print(sales.skew())
sns.distplot(sales)
